VSB.jl Documentation

VSB.jl Documentation

Vortex Sheet Boundary (VSB) Method Documentation

Theory for VSB.jl

Introduction and Motivation

Mathematical Derivation

Another Section

VSB.jl Documentation

VSB.Boundary — Type.

Boundary(body_pts,t_hats,n_hats)

Struct containing boundary information. The user must supply the location of the body points (body_pts), and the tangent and normal vectors of the body at each body point.

VSB.func — Method.

func(x)

Returns double the number x plus 1.